Cloud-Run-Function/requirements.txt
functions-framework>=3.0.0
google-cloud-videointelligence>=2.8.0